gameserver

+++ server/CenterServer/CenterServer.cpp
@@ -42,7 +42,7 @@ void CenterServer::oninit()
 	listenHttp(httpIP, httpPort);
 	
 	unsigned long int timeid = m_Timer.addTimer( 2, -1 );
-	printf("centerserver on oninit tid=%d   id=%d\r\n",pthread_self(), timeid);
+	printf("centerserver on oninit tid=%d   id=%ld\r\n",pthread_self(), timeid);
 }
 
 void CenterServer::onuninit()
@@ -50,12 +50,12 @@ void CenterServer::onuninit()
 
 }
 
-void CenterServer::handleTimerMsg( unsigned int id )
+void CenterServer::handleTimerMsg( unsigned long int id )
 {
 	printf("centerserver on handleTimerMsg tid=%d   id=%d\r\n",pthread_self(), id);
 }
 
-void CenterServer::onMsg( unsigned int id, KernalNetWorkType netType, KernalMessageType type, const char *data, unsigned int size )
+void CenterServer::onMsg( unsigned long int id, KernalNetWorkType netType, KernalMessageType type, const char *data, unsigned int size )
 {
 	printf("centerserver on msg tid=%d id=%d\r\n",pthread_self(), id);
 	m_Locker.lock();
